+++ src/cmd/validate.rs
@@ -6,6 +6,7 @@ use anyhow::{anyhow, Result};
 use csv::ByteRecord;
 #[cfg(any(feature = "full", feature = "lite"))]
 use indicatif::{ProgressBar, ProgressDrawTarget};
+use itertools::Itertools;
 use jsonschema::paths::PathChunk;
 use jsonschema::{output::BasicOutput, JSONSchema};
 #[allow(unused_imports)]
@@ -237,6 +238,7 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
     let mut record = csv::ByteRecord::new();
     // reuse batch buffer
     let mut batch = Vec::with_capacity(BATCH_SIZE);
+    let mut validation_results = Vec::with_capacity(BATCH_SIZE);
     let mut valid_flags: Vec<bool> = Vec::with_capacity(record_count as usize);
     let mut validation_error_messages: Vec<String> = Vec::with_capacity(50);
 
@@ -277,7 +279,7 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
 
         // do actual validation via Rayon parallel iterator
         // validation_results vector should have same row count and in same order as input CSV
-        let validation_results: Vec<Option<String>> = batch
+        batch
             .par_iter()
             .map(|record| {
                 do_json_validation(
@@ -288,21 +290,18 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
                     &schema_compiled,
                 )
             })
-            .collect();
+            .collect_into_vec(&mut validation_results);
 
         // write to validation error report, but keep Vec<bool> to gen valid/invalid files later
         // because Rayon collect() guaranteeds original order, can sequentially append results to vector with each batch
-        for result in validation_results {
-            match result {
-                Some(validation_error_msg) => {
-                    invalid_count += 1;
-                    valid_flags.push(false);
+        for result in &validation_results {
+            if let Some(validation_error_msg) = result {
+                invalid_count += 1;
+                valid_flags.push(false);
 
-                    validation_error_messages.push(validation_error_msg);
-                }
-                None => {
-                    valid_flags.push(true);
-                }
+                validation_error_messages.push(validation_error_msg.to_string());
+            } else {
+                valid_flags.push(true);
             }
         }
 
@@ -449,7 +448,6 @@ fn write_error_report(input_path: &str, validation_error_messages: Vec<String>) 
 }
 
 /// if given record is valid, return None, otherwise, error file entry string
-#[inline]
 fn do_json_validation(
     headers: &ByteRecord,
     headers_len: usize,
@@ -472,7 +470,6 @@ fn do_json_validation(
         schema_compiled,
     )
     .map(|validation_errors| {
-        use itertools::Itertools;
         // squash multiple errors into one long String with linebreaks
         let combined_errors: String = validation_errors
             .iter()